+++ packages/utils/plugins-helpers/src/index.ts
@@ -3,7 +3,6 @@ export * from './types.js';
 export * from './utils.js';
 export * from './helpers.js';
 export * from './federation.js';
-export * from './errors.js';
 export * from './getCachedDocumentNodeFromSchema.js';
 export * from './oldVisit.js';
 export * from './profiler.js';