+++ packages/plugins/java/kotlin/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @graphql-codegen/kotlin
 
+## 2.3.6
+
+### Patch Changes
+
+- Updated dependencies [[`63dc8f205`](https://github.com/dotansimha/graphql-code-generator/commit/63dc8f2054e27b944f7d8dc59db8afa85760a127)]:
+  - @graphql-codegen/visitor-plugin-common@2.13.1
+  - @graphql-codegen/plugin-helpers@2.7.2
+  - @graphql-codegen/java-common@2.2.6
+
+## 2.3.5
+
+### Patch Changes
+
+- Updated dependencies [[`a46b8d99c`](https://github.com/dotansimha/graphql-code-generator/commit/a46b8d99c797283d773ec14163c62be9c84d4c2b)]:
+  - @graphql-codegen/visitor-plugin-common@2.13.0
+  - @graphql-codegen/java-common@2.2.5
+
 ## 2.3.4
 
 ### Patch Changes